+++ src/main/java/operator/AddOperator.java
@@ -2,7 +2,7 @@ package operator;
 
 public class AddOperator implements Operator {
 	@Override
-	public double operate(int a, int b) {
+	public double operate(double a, double b) {
 		return a + b;
 	}
 }
