+++ src/main/java/operator/DivideOperator.java
@@ -2,7 +2,7 @@ package operator;
 
 public class DivideOperator implements Operator {
 	@Override
-	public double operate(int a, int b) {
+	public double operate(double a, double b) {
 		return (double)a / b;
 	}
 }
