+++ src/main/java/operator/MultiplyOperator.java
@@ -2,7 +2,7 @@ package operator;
 
 public class MultiplyOperator implements Operator {
 	@Override
-	public double operate(int a, int b) {
+	public double operate(double a, double b) {
 		return a * b;
 	}
 }
