+++ src/main/java/operator/Operator.java
@@ -1,5 +1,5 @@
 package operator;
 
 public interface Operator {
-	double operate(int a, int b);
+	double operate(double a, double b);
 }
