+++ src/main/java/operator/SubtractOperator.java
@@ -2,7 +2,7 @@ package operator;
 
 public class SubtractOperator implements Operator {
 	@Override
-	public double operate(int a, int b) {
+	public double operate(double a, double b) {
 		return a - b;
 	}
 }
